+++ ubo_app/services/0-camera/ubo_handle.py
@@ -1,10 +1,9 @@
 # ruff: noqa: D100, D101, D102, D103, D104, D107, N999
-from pathlib import Path
-
 from reducer import reducer
 from setup import init_service
 
 from ubo_app.load_services import register_service
+from ubo_app.utils import IS_RPI
 
 register_service(
     service_id='camera',
@@ -12,6 +11,5 @@ register_service(
     reducer=reducer,
 )
 
-IS_RPI = Path('/etc/rpi-issue').exists()
 if IS_RPI:
     init_service()