+++ ubo_app/services/0-keypad/ubo_handle.py
@@ -1,15 +1,13 @@
 # ruff: noqa: D100, D101, D102, D103, D104, D107, N999
-from pathlib import Path
-
 from setup import init_service
 
 from ubo_app.load_services import register_service
+from ubo_app.utils import IS_RPI
 
 register_service(
     service_id='keypad',
     label='Keypad',
 )
 
-IS_RPI = Path('/etc/rpi-issue').exists()
 if IS_RPI:
     init_service()